Peter

+++ src/app/Components/cart/cart.component.ts
@@ -1,27 +1,37 @@
+
 import { ToastrService } from 'ngx-toastr';
-import { IProduct } from '../../Core/interfaces/iproduct';
 import { CartService } from './../../Core/Services/cart.service';
 import { Component, OnInit } from '@angular/core';
-import { SpinnerComponent } from '../spinner/spinner.component';
+import { SpinnerComponent } from '../Layout/spinner/spinner.component';
 import { RouterLink } from '@angular/router';
+import { SharedService } from '../../Core/Services/shared.service';
+import Swal from 'sweetalert2';
+import { ICart } from '../../Core/interfaces/icart';
+import { ICartItem } from '../../Core/interfaces/icart-item';
 
 @Component({
   selector: 'app-cart',
   standalone: true,
   imports: [SpinnerComponent, RouterLink],
   templateUrl: './cart.component.html',
-  styleUrl: './cart.component.css'
+  styles: ''
 })
 export class CartComponent implements OnInit {
   isLoading: boolean = true
-  cartItems: any
+  cartProducts: any
+
+  cart: ICart = {} as ICart
+  items: ICartItem[] = []
+  productQty: number = 1;
+
 
   userId: number = 0
 
 
   constructor(
     private _CartService: CartService,
-    private _ToastrService: ToastrService
+    private _ToastrService: ToastrService,
+    private _SharedService: SharedService
   ) { }
 
   ngOnInit(): void {
@@ -30,9 +40,26 @@ export class CartComponent implements OnInit {
     this.userId = Number(localStorage.getItem("userId"))
 
     //get all items in cart
-    this._CartService.getCartByUserId(this.userId).subscribe({
+    this._CartService.getcartByUserId(this.userId).subscribe({
       next: res => {
-        this.cartItems = res
+        // this.cartItems = res.cartItems; console.log(this.cartItems);
+        this.cart = res;
+        this.items = res.cartItems
+
+        console.log(this.cart);
+        console.log(this.items);
+
+
+      },
+      error: err => console.log(err)
+    })
+
+    //get all products in cart
+    this._CartService.getProductsByUserId(this.userId).subscribe({
+      next: res => {
+        this.cartProducts = res
+        // setTimeout(() => { this._SharedService.updateCartCount(this.cartItems.length) }, 100);
+        this._SharedService.updateCartCount(this.cartProducts.length)
         this.isLoading = false
       }
       ,
@@ -42,15 +69,78 @@ export class CartComponent implements OnInit {
       }
     })
 
+    this._SharedService.cartProdQty.subscribe(res => this.productQty = res)
+
+  }
+
+  incQty(prodId: number) {
+    for (let i = 0; i < this.items.length; i++) {
+      if (this.items[i].productId == prodId) {
+        this._CartService.incQty(this.items[i].id).subscribe({
+          next: () => {
+            this._ToastrService.success("item increased")
+            this.items[i].quantity++
+          },
+          error: err => this._ToastrService.warning(JSON.stringify(err.error))
+        })
+      }
+
+    }
+  }
+
+  decQty(prodId: number) {
+    for (let i = 0; i < this.items.length; i++) {
+      if (this.items[i].productId == prodId) {
+        this._CartService.decQty(this.items[i].id).subscribe({
+          next: () => {
+            this._ToastrService.success("item decreased");
+            this.items[i].quantity--
+          },
+          error: err => this._ToastrService.warning(JSON.stringify(err.error))
+        })
+      }
+
+    }
+
   }
 
   deleteItem(productId: number) {
-    this._CartService.deleteItem(this.userId, productId).subscribe({
-      next: () => {
-        this._ToastrService.show("Item deleted from cart")
-        this.cartItems = this.cartItems.filter((item: any) => item.id != productId)
-      },
-      error: err => this._ToastrService.warning(err)
-    })
+    Swal.fire({
+      title: 'Are You Sure?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Yes, Delete Item',
+      cancelButtonText: 'Cancel'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this._CartService.deleteItem(this.userId, productId).subscribe({
+          next: () => {
+            this._ToastrService.show("Item deleted from cart")
+            this.items = this.items.filter((item: any) => item.productId != productId)
+            Swal.fire(
+              'Done',
+              'item Deleted Successfuly',
+              'success'
+            );
+          },
+          error: err => {
+            Swal.fire(
+              'Wrong!',
+              'something wrong happened in process',
+              'error'
+            );
+            this._ToastrService.warning(err)
+
+          }
+        })
+      }
+    });
+
+
+
   }
+
+
 }
